@@ -137,56 +138,53 @@
 		"xev": "3.0.2"
 	},
 	"devDependencies": {
-		"@redocly/openapi-core": "1.0.0-beta.108",
+		"@redocly/openapi-core": "1.0.0-beta.123",
 		"@types/bcryptjs": "2.4.2",
-		"@types/bull": "3.15.9",
-		"@types/cbor": "6.0.0",
 		"@types/escape-regexp": "0.0.1",
-		"@types/fluent-ffmpeg": "2.1.20",
+		"@types/fluent-ffmpeg": "2.1.21",
 		"@types/js-yaml": "4.0.5",
-		"@types/jsdom": "20.0.0",
-		"@types/jsonld": "1.5.6",
-		"@types/jsrsasign": "10.5.2",
+		"@types/jsdom": "21.1.0",
+		"@types/jsonld": "1.5.8",
+		"@types/jsrsasign": "10.5.5",
 		"@types/koa": "2.13.5",
-		"@types/koa-bodyparser": "4.3.7",
+		"@types/koa-bodyparser": "4.3.10",
 		"@types/koa-cors": "0.0.2",
 		"@types/koa-favicon": "2.0.21",
 		"@types/koa-logger": "3.1.2",
-		"@types/koa-mount": "4.0.1",
+		"@types/koa-mount": "4.0.2",
 		"@types/koa-send": "4.1.3",
-		"@types/koa-views": "7.0.0",
-		"@types/koa__cors": "3.1.1",
+		"@types/koa__cors": "3.3.1",
 		"@types/koa__multer": "2.0.4",
-		"@types/koa__router": "8.0.11",
-		"@types/mocha": "9.1.1",
-		"@types/node": "18.7.16",
-		"@types/node-fetch": "3.0.3",
-		"@types/nodemailer": "6.4.5",
+		"@types/koa__router": "12.0.0",
+		"@types/mocha": "10.0.1",
+		"@types/node": "18.14.1",
+		"@types/node-fetch": "2.6.2",
+		"@types/nodemailer": "6.4.7",
 		"@types/oauth": "0.9.1",
 		"@types/pug": "2.0.6",
 		"@types/punycode": "2.1.0",
 		"@types/qrcode": "1.5.0",
+		"@types/qs": "6.9.7",
 		"@types/random-seed": "0.3.3",
-		"@types/ratelimiter": "3.4.3",
-		"@types/redis": "4.0.11",
+		"@types/ratelimiter": "3.4.4",
 		"@types/rename": "1.0.4",
-		"@types/sanitize-html": "2.6.2",
-		"@types/semver": "7.3.12",
-		"@types/sharp": "0.30.5",
+		"@types/sanitize-html": "2.8.0",
+		"@types/semver": "7.3.13",
+		"@types/sharp": "0.31.1",
 		"@types/sinonjs__fake-timers": "8.1.2",
 		"@types/speakeasy": "2.0.7",
 		"@types/tinycolor2": "1.4.3",
 		"@types/tmp": "0.2.3",
-		"@types/uuid": "8.3.4",
+		"@types/uuid": "9.0.1",
 		"@types/web-push": "3.3.2",
 		"@types/websocket": "1.0.5",
-		"@types/ws": "8.5.3",
-		"@typescript-eslint/eslint-plugin": "5.36.2",
-		"@typescript-eslint/parser": "5.36.2",
+		"@types/ws": "8.5.4",
+		"@typescript-eslint/eslint-plugin": "5.53.0",
+		"@typescript-eslint/parser": "5.53.0",
 		"cross-env": "7.0.3",
-		"eslint": "8.23.0",
-		"eslint-plugin-import": "2.26.0",
-		"execa": "6.1.0",
-		"typescript": "4.8.3"
+		"eslint": "8.35.0",
+		"eslint-plugin-import": "2.27.5",
+		"execa": "7.0.0",
+		"typescript": "4.9.5"
 	}
 }
