@@ -12,80 +12,79 @@
 	},
 	"dependencies": {
 		"@discordapp/twemoji": "14.0.2",
-		"@rollup/plugin-alias": "4.0.2",
+		"@rollup/plugin-alias": "4.0.3",
 		"@rollup/plugin-json": "6.0.0",
 		"@rollup/pluginutils": "5.0.2",
 		"@syuilo/aiscript": "0.11.1",
-		"@tabler/icons-webfont": "^2.4.0",
+		"@tabler/icons-webfont": "^2.5.0",
 		"@vitejs/plugin-vue": "4.0.0",
-		"@vue/compiler-sfc": "3.2.45",
+		"@vue/compiler-sfc": "3.2.47",
 		"autobind-decorator": "2.4.0",
-		"autosize": "5.0.1",
-		"blurhash": "1.1.5",
-		"broadcast-channel": "4.14.0",
+		"autosize": "6.0.1",
+		"blurhash": "2.0.5",
+		"broadcast-channel": "4.20.2",
 		"browser-image-resizer": "github:misskey-dev/browser-image-resizer",
-		"chart.js": "3.9.1",
-		"chartjs-adapter-date-fns": "2.0.0",
-		"chartjs-plugin-gradient": "0.5.1",
-		"chartjs-plugin-zoom": "1.2.1",
-		"compare-versions": "5.0.1",
-		"cropperjs": "2.0.0-beta",
-		"date-fns": "2.29.2",
+		"chart.js": "4.2.1",
+		"chartjs-adapter-date-fns": "3.0.0",
+		"chartjs-plugin-gradient": "0.6.1",
+		"chartjs-plugin-zoom": "2.0.0",
+		"compare-versions": "6.0.0-rc.1",
+		"cropperjs": "1.5.13",
+		"date-fns": "2.29.3",
 		"escape-regexp": "0.0.1",
-		"eventemitter3": "4.0.7",
+		"eventemitter3": "5.0.0",
 		"idb-keyval": "6.2.0",
 		"insert-text-at-cursor": "0.3.0",
-		"json5": "2.2.1",
-		"katex": "0.15.6",
-		"matter-js": "0.18.0",
+		"json5": "^2.2.3",
+		"katex": "0.16.4",
+		"matter-js": "0.19.0",
 		"mfm-js": "0.23.3",
-		"misskey-js": "0.0.14",
-		"photoswipe": "5.3.2",
+		"misskey-js": "0.0.15",
+		"photoswipe": "5.3.6",
 		"prismjs": "1.29.0",
-		"punycode": "2.1.1",
-		"querystring": "0.2.1",
+		"punycode": "2.3.0",
 		"rndstr": "1.0.0",
 		"s-age": "1.1.2",
-		"sass": "1.54.9",
+		"sass": "1.58.3",
 		"seedrandom": "3.0.5",
 		"strict-event-emitter-types": "2.0.0",
 		"stringz": "2.1.0",
 		"syuilo-password-strength": "0.0.1",
 		"textarea-caret": "3.1.0",
-		"three": "0.144.0",
+		"three": "0.150.0",
 		"throttle-debounce": "5.0.0",
-		"tinycolor2": "1.4.2",
-		"tsc-alias": "1.7.0",
-		"tsconfig-paths": "4.1.0",
+		"tinycolor2": "1.6.0",
+		"tsc-alias": "1.8.2",
+		"tsconfig-paths": "3.14.2",
 		"twemoji-parser": "14.0.0",
-		"typescript": "4.8.3",
+		"typescript": "4.9.5",
 		"uuid": "9.0.0",
-		"vanilla-tilt": "1.7.2",
-		"vite": "3.1.0",
-		"vue": "3.2.39",
+		"vanilla-tilt": "1.8.0",
+		"vite": "4.1.4",
+		"vue": "3.2.47",
 		"vue-prism-editor": "2.0.0-alpha.2",
 		"vuedraggable": "4.0.1"
 	},
 	"devDependencies": {
 		"@types/escape-regexp": "0.0.1",
-		"@types/glob": "8.0.0",
-		"@types/gulp": "4.0.9",
+		"@types/glob": "8.1.0",
+		"@types/gulp": "4.0.10",
 		"@types/gulp-rename": "2.0.1",
-		"@types/katex": "0.14.0",
-		"@types/matter-js": "0.18.1",
+		"@types/katex": "0.16.0",
+		"@types/matter-js": "0.18.2",
 		"@types/punycode": "2.1.0",
-		"@types/seedrandom": "3.0.2",
+		"@types/seedrandom": "3.0.5",
 		"@types/throttle-debounce": "5.0.0",
 		"@types/tinycolor2": "1.4.3",
-		"@types/uuid": "8.3.4",
-		"@typescript-eslint/eslint-plugin": "5.36.2",
-		"@typescript-eslint/parser": "5.36.2",
+		"@types/uuid": "9.0.1",
+		"@typescript-eslint/eslint-plugin": "5.53.0",
+		"@typescript-eslint/parser": "5.53.0",
 		"cross-env": "7.0.3",
-		"cypress": "10.7.0",
-		"eslint": "8.23.0",
-		"eslint-plugin-import": "2.26.0",
-		"eslint-plugin-vue": "9.4.0",
-		"rollup": "2.79.0",
-		"start-server-and-test": "1.14.0"
+		"cypress": "12.7.0",
+		"eslint": "8.35.0",
+		"eslint-plugin-import": "2.27.5",
+		"eslint-plugin-vue": "9.9.0",
+		"rollup": "3.17.3",
+		"start-server-and-test": "1.15.4"
 	}
 }
