@@ -1,18 +1,19 @@
 {
 	"name": "sw",
 	"private": true,
+	"type": "module",
 	"scripts": {
 		"watch": "node build.js watch",
 		"build": "node build.js",
-		"lint": "eslint --quiet src/**/*.{ts}"
+		"lint": "eslint --quiet src/**/*.ts"
 	},
 	"resolutions": {},
 	"dependencies": {
-		"esbuild": "^0.14.13",
-		"idb-keyval": "^6.0.3",
-		"misskey-js": "0.0.14"
+		"esbuild": "^0.17.10",
+		"idb-keyval": "^6.2.0",
+		"misskey-js": "0.0.15"
 	},
 	"devDependencies": {
-		"eslint": "^8.2.0"
+		"eslint": "^8.35.0"
 	}
 }
