@@ -1,5 +1,9 @@
-const fs = require('fs');
-const meta = require('../package.json');
+import { mkdirSync, writeFileSync } from 'fs';
+import * as meta from '../package.json' assert { type: "json" };
+import path from 'path';
+import url from 'url';
 
-fs.mkdirSync(__dirname + '/../built', { recursive: true });
-fs.writeFileSync(__dirname + '/../built/meta.json', JSON.stringify({ version: meta.version }), 'utf-8');
+const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
+
+mkdirSync(__dirname + '/../built', { recursive: true });
+writeFileSync(__dirname + '/../built/meta.json', JSON.stringify({ version: meta.version }), 'utf-8');
